Close #2
Adds handling of when certain save files have not been created yet. The
bot keeps running and replies to the user instead of exiting and throwing
the errors. Also adds other notifications to the user for certain events
e.g. if they try to start the poll with no options.